Revert "float input keeps crashing the fw so switching to using uint32_t for currents"
This reverts commit 7557bfba4e6abd5bc8518b3a0716dfe891c77967.

+++ stm32-modules/include/flex-stacker/flex-stacker/messages.hpp
@@ -92,8 +92,8 @@ struct ForceUSBDisconnect {
 struct SetMotorCurrentMessage {
     uint32_t id;
     MotorID motor_id;
-    uint32_t run_current;
-    uint32_t hold_current;
+    float run_current;
+    float hold_current;
 };
 
 struct SetTMCRegisterMessage {
